requirements.txt
<<<<<<< HEAD
# ======================================================
# 🌱 EcoPackAI – Backend Frameworks
# ======================================================
fastapi==0.110.0
uvicorn[standard]==0.27.1
starlette==0.36.3

Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0

python-multipart==0.0.9

# ======================================================
# 🔐 Authentication & Security
# ======================================================
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# ======================================================
# 🌐 OAuth (Google Login)
# ======================================================
google-auth==2.27.0
google-auth-oauthlib==1.1.0
httpx==0.26.0

# ======================================================
# 🗄️ Database & ORM
# ======================================================
SQLAlchemy==2.0.25
pymysql==1.1.0
mysql-connector-python=9.4.0
alembic==1.13.1

# ======================================================
# ⚙️ Configuration & Environment
# ======================================================
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# ======================================================
# 📊 Data Processing (ML DEPENDENCY – DO NOT TOUCH)
# ======================================================
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
openpyxl==3.1.2

# ======================================================
# 🤖 Machine Learning (DO NOT CHANGE – PRECIOUS)
# ======================================================
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
scipy==1.11.1
shap==0.44.0
rapidfuzz==3.6.1

# ======================================================
# 📈 Visualization / BI
# ======================================================
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ======================================================
# 🖥️ Streamlit Frontend
# ======================================================
streamlit==1.31.0
requests==2.31.0

# ======================================================
# 📄 PDF & Report Generation
# ======================================================
=======
# ======================================================
# 🌱 EcoPackAI – Backend Frameworks
# ======================================================
fastapi==0.110.0
uvicorn[standard]==0.27.1
starlette==0.36.3

Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0

python-multipart==0.0.9

# ======================================================
# 🔐 Authentication & Security
# ======================================================
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# ======================================================
# 🌐 OAuth (Google Login)
# ======================================================
google-auth==2.27.0
google-auth-oauthlib==1.1.0
httpx==0.26.0

# ======================================================
# 🗄️ Database & ORM
# ======================================================
SQLAlchemy==2.0.25
pymysql==1.1.0
mysql-connector-python=9.4.0
alembic==1.13.1

# ======================================================
# ⚙️ Configuration & Environment
# ======================================================
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# ======================================================
# 📊 Data Processing (ML DEPENDENCY – DO NOT TOUCH)
# ======================================================
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
openpyxl==3.1.2

# ======================================================
# 🤖 Machine Learning (DO NOT CHANGE – PRECIOUS)
# ======================================================
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
scipy==1.11.1
shap==0.44.0
rapidfuzz==3.6.1

# ======================================================
# 📈 Visualization / BI
# ======================================================
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ======================================================
# 🖥️ Streamlit Frontend
# ======================================================
streamlit==1.31.0
requests==2.31.0

# ======================================================
# 📄 PDF & Report Generation
# ======================================================
>>>>>>> a610f336bdcd0e8d457dac84df0e7c5a83fc5688
reportlab==4.0.9